Update TenantService to support saving tenant from booking and enhance fee handling
- Add `saveTenantOrFromBooking` method in `TenantService` to handle tenant creation from booking.
- Update `priceInput` type to `BigDecimal` in `TenantOtherFee` and `OtherFeeDTO`.
- Enhance fee calculation methods in `TenantBillGenService` to use `BigDecimal` for better precision.
- Add necessary imports and update related logic.

+++ homi-model/src/main/java/com/homi/model/dao/entity/TenantOtherFee.java
@@ -8,6 +8,7 @@ import lombok.ToString;
 
 import java.io.Serial;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * <p>
@@ -52,7 +53,7 @@ public class TenantOtherFee implements Serializable {
 
     @Schema(description = "价格输入值")
     @TableField("price_input")
-    private Integer priceInput;
+    private BigDecimal priceInput;
 
     @Schema(description = "是否删除：0=否，1=是")
     @TableField("deleted")
